X-VLA finetune to a bundle the robot actually launches
First X-VLA fine-tune on the excavator (masi_digging_dry_2, chunk 30, 10k steps,
bf16, 2h49m) and the four things between a trained checkpoint and a machine that
drives on it. Each was a refusal at launch on the Jetson, found one at a time, so
each fix is also a check that fails at EXPORT time instead.

export_split_onnx: record the instruction strings and the control rate.
`xvla_split.resolve_tasks` and `resolve_fps` take both from bundle.json or refuse
to start -- rightly, since the policy conditions on the language embedding and the
chunk is rate commands, so a wrong instruction or a wrong rate is silently wrong
motion rather than an error. The exporter wrote neither. Tasks come from the
dataset's tasks.parquet in task_index order (the whole LIST for a multi-task
finetune, both `task` and `tasks` so an older runtime still starts), fps from
meta/info.json. Note the asymmetry: fps resolves from the export venv but tasks do
not, because .venv-export carries neither pandas nor pyarrow and cannot read
parquet at all -- pass --task explicitly from there. That failure now prints why;
the first version swallowed it in a bare except and shipped an unlaunchable bundle.

export_split_onnx: size_mb from the file on disk, not from _meta scratch. A bundle
rebuilt over rewritten graphs -- which is exactly what tools/fp16_weights.py
produces -- otherwise misreports its own footprint by 2x, and that is the number
you consult to ask whether it fits in 8 GB.

parity.py: upcast the reference to fp32. run_digging.sh trains bf16 by default and
export_split_onnx traces through .eval().float(), so the graphs are fp32 upcast
from bf16 weights. Against a bf16 reference this raised "expected m1 and m2 to
have the same dtype"; the worse outcome was the one where it did not raise and
measured bf16 rounding as export error. With the reference upcast identically the
finetune passes at cos 1.000000, max|d| 0.00000 on the physical action -- the
first parity number for an X-VLA finetune rather than the base.

retask_dataset.py: rename an instruction across a LeRobot v3 dataset. The
recordings disagree on phrasing for one job -- "move the sand to the container" vs
"move sand to container" -- and merging them as-is trains two unrelated
instructions on what should be one, splitting the sand data. Metadata-only: the
string lives in meta/tasks.parquet and meta/episodes, never in data/, which carries
task_index. Refuses to rename onto an existing task, since that collapses two
task_index values and changes their meaning in every shard.

Also: dry2_ir wired into the X-VLA runner and preflight; chain scripts that ran the
overnight sequence unattended (train -> score -> export -> X-VLA), both waiting on a
PID or a marker line the job writes, never on a pgrep pattern that can match the
waiter's own argv; and the workstation hardening from earlier the same day
(setup/preflight/smoke/queue entry points, pinned Hub revision) which was still
uncommitted and which the chain depends on.

Verified on the machine: the fp16 bundle built its engines and drove.

+++ orin-nano/xvla-runtime/parity.py
@@ -120,7 +120,15 @@ def emit_reference(args) -> None:
     cfg_obj = PreTrainedConfig.from_pretrained(str(args.checkpoint))
     cfg_obj.device = "cpu"
     ref = XVLAPolicy.from_pretrained(str(args.checkpoint), config=cfg_obj)
-    ref.to("cpu").eval()
+    # float(), to match the exporter exactly. A bf16 fine-tune (run_digging.sh trains with
+    # --policy.dtype=bfloat16 by default) loads with bf16 weights, and export_split_onnx
+    # traces every module through `.eval().float()` -- so the GRAPHS are fp32 upcast from
+    # those weights. Comparing fp32 graphs against a bf16 reference would either fail
+    # outright ("expected m1 and m2 to have the same dtype", which is what a bf16
+    # checkpoint did here) or, worse, quietly measure bf16 rounding and report it as
+    # export error. Upcasting the reference the same way makes the two paths the same
+    # numerics, so what is left is genuinely the split's own error.
+    ref.to("cpu").float().eval()
     model = ref.model
 
     x1 = rng.standard_normal((1, chunk_size, model.dim_action)).astype(np.float32)
